Improving the clarity of the code
Caller of the function(pldm_bios_table_attr_value_entry_enum_decode_handles)
provides the pointer to the buffer and its size.
Modifying the buffer size and using it in the library can be errorprone
Modifying the code to just read the buffer size and not to modify it.

Tested:
Unit Tests passed.

Signed-off-by: Archana Kakani <archana.kakani@ibm.com>