pdr: Add contained entity to an association PDR
Include function to add a contained entity into an entity association
PDR. A contained entity may be added to an existing container thereby
Imodifying the association PDR or it can be added to the association
PDR by attaching both the container and its contained entity. This API
covers both these cases.

Signed-off-by: Varsha Kaverappa <vkaverap@in.ibm.com>