docs: checklists: Add some more influences and references
Introduce "weird machines" for those who aren't aware of them. The blog
also links to Dullien's paper that provides fairly tight definition
for them.

The SEI CERT C Coding Standard is maintained by members of the C
standards committee as a more approachable resource for C programmers
than the standard itself (which is written for implementers of C more so
than users).

Change-Id: I6df09feacd47c0eea42b840784ab2c90a913fcff
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>