pdr: PDR loop fix
While creating the Entity association PDRs, same record handle
is used multiple time to create PDR. Because of which 2 PDRs
with same record handle are added to the repo causing loop.
Adding the fix to resolve the duplicate record handle issue.

Tested:
Power on/off and reset reload on mex and non mex systems.

Signed-off-by: Archana Kakani <archana.kakani@ibm.com>